Added StatisticsReportModal
This popup modal will be triggered by the "Statistics Report" button in the ViewData component. It will display a data summary of the selected DataGrid rows. The data summary will consist of all the data analysis listed in the project requirements. It is NOT integrated with ViewData at this point